Adding question frames and quiz functionality
Variables for whether complex number, differentiation and integration questions will be added to the quiz. A new questionPage class, which iterates through the questionPage instances. The rootFrame now has a generate_quiz method, which initializes the questions into the group of frames, each question it's own instance with it's own number and later question.