Integration of Types of Questions
Here, I've added compatibility for more than one list to be input, so that the quiz randomizes between all three dictionaries (differentiation_questions, integration_questions, complex_questions) when sections are turned off. Balancing needs to be done so that the user doesn't get a lot of the same type in a row, there should be a maximum of 3 in each section.