Revise list code + Fix random error
The "generate_quiz" method in rootFrame has been revised, so that the code no longer deletes elements from the new list. This was preventing me from checking if a duplicae question was in the list. I have made code to check whether a difficulty change has been made, for functionality.