Adding User Data and Menu
Menus will be used to see the scoreboard and highscores, and the grades of each user if they want to track. There is a new UserData class which stores each user's data in a dictionary in the root frame by name. This means duplicate names won't work, for now. I've also had to add a new quiz button, slightly different to the restart button which only restarts the quiz for the current user, instead of making a new one. Right now, everything stores correctly although it's difficult to see.