Remove themes, add grid
Themes appear to only be compatible with desktops with it installed. Adding this as a feature would not be good since it requires extra steps which are often inaccessible (especially in the case of school computers), so I removed them. Currently I am adding grid configurations for all widgets, some will need individual frames.