Theme-changing change, add number_correct
Themes are now changed in a sub-menu of the file menu. Later I will change this to an options menu. There is no longer an options popup, but the application still needs to be restarted. I've also added the number_correct parameter so that users can check the raw input of how many they got, for both scoreboards and the end of the quiz.