Add a load user feature (+ reset all users)
Users are put into the load user menu in filemenu. It only stores user names currently, and functionality is incomplete when resetting users or restarting. However, I might remove this feature and replace it with a load-theme button, allowing users to customize how the quiz looks by themes instead of users, since there isn't really a point to loading users (as it will just be overwritten anyways).