Add Checkbuttons + Save user feature
Checkbuttons are generated on the side of the quiz based on how many questions there are. If the user gets it right, it ticks, otherwise it stays off. This will be changed so that if the user chooses an answer regardless of if it's right or wrong, it will mark it as done, but if they skip it won't. I will also change checkbuttons to labels.
The save user feature has been implemented so that the user is saved when the quiz is done instead.